api/requirements.txt
requests
fastapi
psycopg2-binary
schedule
uvicorn